Route protection, password hashing. API Termination

+++ app/application/service/client_service.py
@@ -4,6 +4,9 @@ from app.domine.models.client import Client
 from fastapi import HTTPException
 from typing import List, Optional
 from uuid import UUID
+from passlib.context import CryptContext
+
+pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
 
 class ClientService:
     def __init__(self, repository: ClientRepository):
@@ -11,18 +14,28 @@ class ClientService:
 
     def create_client(self, create_client: ClientCreate) -> ClientResponse:
         existing_client = self.repository.get_client_with_username(create_client.username)
+        hashed_password = pwd_context.hash(create_client.password)
 
         if existing_client:
-            raise HTTPException(status_code=400, detail=f'El usuario ya existe')
+            raise HTTPException(status_code=400, detail='The user already exists')
         
-        client = Client(**create_client.model_dump())
-        created = self.repository.create_client(client)
-        return ClientResponse.model_validate(created)
+        client = Client(
+            username=create_client.username,
+            password=hashed_password,
+            email=create_client.email,
+            address=create_client.address,
+            phone=create_client.phone,
+            role_id=2
+        )
+
+        created_client = self.repository.create_client(client)
+        return ClientResponse.model_validate(created_client)
             
     def get_all_clients(self) -> List[ClientResponse]:
         clients = self.repository.get_all_clients()
-        if not clients:
-            raise HTTPException(status_code=404, detail='No existen clientes')
+
+        if clients == []:
+            raise HTTPException(status_code=404, detail='There are no clients')
         
         list_clients = []
         
@@ -36,7 +49,7 @@ class ClientService:
 
         if client_find is None:
             raise HTTPException(
-                status_code=404, detail=f'Cliente no encontrado'
+                status_code=404, detail=f'Client {username} not found'
             )
         
         return ClientResponse.model_validate(client_find)
@@ -46,8 +59,17 @@ class ClientService:
         
         if client_find is None:
             raise HTTPException(
-                status_code=400, detail='El cliente no se pudo actualizar'
+                status_code=404, detail='The client could not be updated'
             )
+        
+        username_exists = self.repository.get_client_with_username(update_client.username)
+        if username_exists and username_exists.id != client_find.id:
+            raise HTTPException(status_code=400, detail=f'El usuario {update_client.username} ya existe')
+        
+
+        password_exists = self.repository.get_client_with_password(update_client.password)
+        if password_exists and password_exists.id != client_find.id:
+            raise HTTPException(status_code=400, detail='A user with that password already exists.')
         
         client_find.username = update_client.username
         client_find.email = update_client.email
@@ -60,9 +82,11 @@ class ClientService:
     def delete_client(self, username: str) -> ClientResponse:
         client = self.repository.get_client_with_username(username)
 
+        if client.is_active == False:
+            raise HTTPException(status_code=400, detail=f'The user {client.username} is already inactive')
         if client is None:
             raise HTTPException(
-            status_code=400, detail=f'El cliente no se encontró'
+            status_code=400, detail=f'The client {username} was not found'
         )
 
         deleted_client = self.repository.delete_client(client)
